Move to src
Wanted to see if I could get this to use src as the base directory.
The answer is yes.
Tweaked whitespace too.
