--- Calculator/CalculatorContent/Main.qml
@@ -1,7 +1,0 @@
-import Calculator
-
-MainUI {
-    anchors.fill: parent
-
-    display.text: backend ? backend.display : ""
-}
