+++ src/App/autogen/environment.h
@@ -27,10 +27,10 @@ Q_IMPORT_QML_PLUGIN(QtQuick_Studio_EventSystemPlugin)
 #endif
 
 inline void set_qt_environment() {
-    qputenv("QML_COMPAT_RESOLVE_URLS_ON_ASSIGNMENT", "1");
-    qputenv("QT_ENABLE_HIGHDPI_SCALING", "0");
-    qputenv("QT_LOGGING_RULES", "qt.qml.connections=false");
-    qputenv("QT_QUICK_CONTROLS_CONF", ":/qtquickcontrols2.conf");
+	qputenv("QML_COMPAT_RESOLVE_URLS_ON_ASSIGNMENT", "1");
+	qputenv("QT_ENABLE_HIGHDPI_SCALING", "0");
+	qputenv("QT_LOGGING_RULES", "qt.qml.connections=false");
+	qputenv("QT_QUICK_CONTROLS_CONF", ":/qtquickcontrols2.conf");
 }
 
 #endif // APP_ENVIRONMENT_H
