+++ src/CMakeLists.txt.shared
