+++ src/CMakeLists.txt.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 17.0.1, 2025-10-01T00:11:33. -->
+<!-- Written by QtCreator 17.0.1, 2025-10-01T01:31:07. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
