+++ src/Calculator/designer/plugin.metainfo
