+++ src/CalculatorContent/Main.qml
@@ -1,0 +1,7 @@
+import Calculator
+
+MainUI {
+	anchors.fill: parent
+
+	display.text: backend ? backend.display : ""
+}
