+++ src/qtquickcontrols2.conf
